dracut: Make sure the kernel modules are installed, no matter what
The behavior of dracut, in `--host-only` mode, is to NOT include the
modules that are not loaded in the initrd, even when they're explicitly
installed in the `installkernel()` function with `instmods`.

This comes a bit as a surprise, but that's how it works.

This can be seen by enabling `--debug` when calling dracut-install:

    dracut-install: Handle module 'nls_ascii'
    dracut-install: dracut_install 'nls_ascii' not hostonly

Which corresponds to this part of dracut-install.c:

        if (arg_hostonly && !check_hashmap(modules_loaded, name)) {
                log_debug("dracut_install '%s' not hostonly", name);
                return 0;
        }

However, we want to be 100% sure that the modules needed to mount the
ESP are included.

Looking at other modules, it looks like the common way to force modules
to be installed is to set `hostonly=` in the environment before callin
instmods.

This commit does just that.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>