Introduce the SteamOS offloading scheme
The SteamOS offloading scheme allows to offload some directories to
/home. These are directories that might grow big, bigger than what /var
allows, so they need to live on the "big" SteamOS partition: /home.

Some of these directories are NOT on the var partition (/opt, /srv), so
in this case the offloading makes sure these directories are writable
(in case the rootfs is read-only), additionally to making sure that they
can grow big.

For the backing directory, we chose to use `/home/.steamos/offload`.
The idea is that we might need to do more SteamOS specific trickery in
the future, and in this case we'll be happy to be able to use
`/home/.steamos/whatever-hack` for that purpose.

The offloading scheme is implemented with systemd bind mount units, that
are grouped together under `steamos-offload.target`. Having a target
makes checking for success or failure of the offloading scheme as a
whole convenient, ie. `systemctl status steamos-offload.target`.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>