dracut: Rewrite the steamos-etc-overlay hook
In SteamOS we need to mount the /etc overlay during the initrd. This is
done in the dracut hook "steamos-etc-overlay hook.sh", along with the
initial setup: creating the directories needed for the overlay (upper
and work dir).

In order to do that, we used to read the /etc/fstab, find the line
corresponding to the etc overlay, parse it, and then use it to create
the {upper,work} dir and then mount the overlay.

Parsing fstab is not great, but since the file /etc/fstab was defined in
another git repository, I prefered to not pretend to know what was
written in it, and I found that parsing it was (arguably) more robust.

However, all of that is now obsolete. In the previous commit, we moved
the definition of the etc overlay out of the /etc/fstab file, and
prefered to use a systemd mount unit instead.

This change obviously breaks the steamos-etc-overlay hook, so let's
rewrite it.

In this re-write, I prefer to avoid parsing anything. Now that the
overlay is defined in the same git repository, it's easier to maintain
both the etc mount unit and the dracut hook together.

NOTE: mounting the overlay manually from the initrd doesn't change
anything regarding the existing etc.mount unit. Once the systemd is
booted, systemd manages this mount point, `systemctl status etc.mount`
shows the unit as enabled, and on shutdown systemd will automatically
unmount the etc overlay.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>