misc: steamos-clone: Handle new log directory location (now on /home)
In recent commits, we started to offload /var/log to /home, but since we
want to keep two log directories (log-A and log-B), it becomes a bit of
a special case that needs to be handled with care.

This commit takes care of that, ie. don't copy the logs of the current
OS when we install it to another device.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>