offload: generator: Expose log directories in /run/steamos/mnt
The idea is that other parts of SteamOS might need to know and access
the log directories for 'self' or 'other', so let's expose it here, once
and for all, to avoid that various tools come up with their own ways of
mapping var-{A,B} to var-{self,other}.

Following the logic of the previous commit that bind-mounted the right
log directory over /var/log, this commit creates more bind-mounts under
a well-known name: /run/steamos/mnt/log-{self,other}.

Note that the use of 'self' and 'other' follows the terminology already
used in /dev/disk/steamos.

Using /run/steamos/mnt might or might not be the best idea ever. But the
idea is that these directories are only needed for very specific SteamOS
parts (installer, atomic updater, and probably nothing else), so I don't
want to expose it "publicly", like in `/var/log-{self,other}`.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>