rauc: post-install: Handle new log directory location (now on /home)
In recent commits, we started to offload /var/log to /home, but since we
want to keep two log directories (log-A and log-B), it becomes a bit of
a special case that needs to be handled with care.

The first change is that we don't need to handle /var/log as a
special-case when we sync the two var partitions, as now /var/log is a
bind-mount, meaning it's automatically excluded from the synchronization
thanks to the option --one-file-system.

The second change is that now, we take care of cleaning the logs of
the 'other' SteamOS after and update was performed. This is because
these "old logs" have no relation with the new SteamOS that was just
installed, so keeping them around would just make things confusing.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>