atomic-update: rauc: post-install: Keep var/log when syncing var partitions
My initial assumption was that logs should not be copied from A to B
during an upgrade, and a newly installed rootfs should have no previous
logs, because, well, it's new and was never been booted yet.

However, for debugging boot issues after an update, having all the logs
around might actually be very convenient.

When people do snapshot/rollback of their root filesystem with btrfs,
/var/log is one of the first things they suggest splitting out into a
submodule so that it is unaffected by rollbacks. If something goes
horribly wrong after an upgrade, surely we want the logs from when it
went wrong to be "in continuity" with the logs from before the bad
upgrade, and the logs from after we rolled back?

To illustrate that with an example: we would expect the Journal/syslog
or whatever to be like this (the "A" and "B" in the first column
indicate which partition was booted at the time and are not going to be
in the real logs):

  A 10:00 steamos: booting version 100 from partition A
  A 10:01 steamos: ...
  A 10:02 steamos: downloading version 101 to partition B...
  A 10:03 steamos: rebooting to upgrade to version 101 on partition B
  B 10:04 steamos: booting version 101 from partition B
  B 10:05 steamos: ...
  B 10:06 steamos: Segmentation fault, core dumped
  B 10:09 steamos: Bad things have happened, rebooting into partition A to recover
  A 10:10 steamos: booting version 100 from partition A
  A 10:11 steamos: ...
  A 11:58 steamos: downloading version 102 to partition B...
  A 11:59 steamos: rebooting to upgrade to version 102 on partition B
  B 12:00 steamos: booting version 102 from partition B
  B 12:01 steamos: ...

So this commit simply stops excluding /var/log from the sync, which
means that logs are now carried around during an upgrade from A to B.

This is only part of the solution, because in case of rollback, we lose
the logs from the system that failed to boot. See next commit for the
complete solution.

(thx smcv)

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>