offload: Add var/log to the list of offloaded directories
With this commit, /var/log is indeed shared between A and B, meaning
that we can have one linear log history, independently from the fact
that we boot A or B, or even rollback.

The additional benefit is that we don't risk to fill up the var
partition with logs, in case some process goes horribly off track and
spits out gigabytes of logs,

See previous commit for a longer explanation.

(thx smcv)

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>