misc: steamos-clone: Stop caring about /var/log
Now that /var/log is offloaded to /home, there's not much need to remove
it from the var partition on the target device.

To be consistent, if we still wanted to remove the logs, we should now
remove the directory $offload_dir/var/log. However, maybe let's just
keep things simple, and simply not care and let it be.

One of the argument of keeping the logs is to ease debugging boot
issues, and it still applies here. If you install SteamOS to a device
with steamos-clone, and then it doesn't boot, it can be helpful to have
the previous logs around.

So this commit simplify things by NOT trying to handle /var/log in a
smart way, which in the end is not smart at all.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>