misc: Add steamos.sh profile
Some executables require to have environment variables set.

For example, the script pacaur (i.e. the pacman for the AUR repository)
needs EDITOR or VISUAL to pop-up the PKGBUILD file and the install
script before the user agrees to install the AUR package.

The scripts located in the directory /etc/profile.d are good places to
set system-wide things related to the distribution.

This commits sets the EDITOR to vim and VISUAL to gvim for every users.
Users may override these variable in their shell profile script.

Fixes:
	$ pacaur -S google-chrome
	:: $VISUAL and $EDITOR environment variables not set or defined editor not found