steamos-glx-driver: Add a helper to get the name of the initrd
With this commit, the name of the initrd is correct for both Debian and
ArchLinux.

Note that we make the assumption that we're not using a lts kernel, just
the normal kernel. Since ArchLinux doesn't name their initrd according
to `uname -r`, it's not obvious how to guess which initrd was booted.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>