steamos-glx-driver: Send notices and warnings to /dev/console
This is a hack.

The attempt is to display those precious messages during the boot (you
know, this screen that is controlled by systemd and display the status
of every services that is started),

There's no obvious way to do that, unless we want to redirect the
outpout of the whole script (in this case, it can be done easily with
`StandardOutput=journal+console`).

Writing to /dev/console is a hack, that actually doesn't really work,
well, some messages are displayed, other are not. It's not great, but at
the moment it's better than nothing, so be it.

With this commit, and if luck is on our side, during boot we can see
the main steps of steamos-glx-driver, ie:

1. build dkms modules
2. build initrd
3. reboot

And it's better than watching the screen with absolutely no feedback on
what's going on.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>