steamos-glx-driver: Rebuild initrd manually after dkms install
The dkms code that looks for the initrd name is:

~~~~
    # $1 = kernel version

    # Find out what the proper filename will be
    for initrd in "initrd-$1.img" "initramfs-$1.img" "initrd.img-$1" "initrd-$1" ''; do
        [[ $initrd && -f $initrd_dir/$initrd ]] && break
    done
    if ! [[ $initrd ]]; then
        # Return if we cannot find an initrd.
        warn $"Unable to find an initial ram disk that I know how to handle." \
            $"Will not try to make an initrd."
        return 0
    fi
~~~~

At the moment our initrd is installed under /boot/initramfs-linux.img,
so it doesn't match.

Note that for now, ArchLinux didn't finish integrating dracut, so the
bits and pieces that install the initrd under this name are our own. We
don't know what name ArchLinux will pickup for the dracut initrd.

Also note that the name for the kernel has been discussed for 10 years
without reaching consencus: https://bugs.archlinux.org/task/16702

So let's just wait for now, and hope the dracut integration will
improve, so that we can use their install script, ie. have the
definitive name for the initrd. From that point we can work on a better
solution.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>