glx: Stop shipping the nvidia.conf dkms snippet
Meaning that DKMS won't try to rebuild the initrd (it doesn't work for
ArchLinux anyway), meaning that we do it ourselves.

I think it's best to stick to the principle of least surprise here, and
follow the ArchLinux behavior.  In ArchLinux world, when the nvidia
drivers are updated, you are expected to rebuild the initrd by yourself.

Let's simplify our life and do the same (ie. do nothing).

If ever it turns out to be a problem in the future, I think we'd rather
try to solve it at the distribution level, maybe discuss with Arch
upstream a better nvidia-dkms integration, rather than using DKMS'
REMAKE_INITRD option.

The way dkms do it is all by "guessing", and it's rather clumsy. DKMS is
just not the right place to do this kind of things.

So this commit does two things:
- we stop shipping a file /etc/dkms/nvidia.conf
- we adjust the code in steamos-glx-driver (mainly, re-check if the
  initrd needs to be rebuilt)
- clarify this point as much as I can with a lengthy comment

This commit also undo db6bcf0e8bdeba31e8a38f2a8ecae5d0f8f06152.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>