grub: Remove traces in early config
The grub early config outputs the following traces:

	---- Entering early config ----
	cmdpath: (hd0,gpt2)/EFI/steamos
	root   : hd0,gpt6
	prefix : (hd0,gpt6)/boot/grub
	-------------------------------
	root partition f11791c4-9448-43e5-b2ae-8d26f4cf1b50 selected to boot
	---- Leaving early config ----
	root   : hd0,gpt6
	prefix : (hd0,gpt6)/boot/grub
	cfg    : (hd0,gpt2)/EFI/steamos/grub-root.cfg
	------------------------------

There traces are for debug purpose only and polute the boot sequence.

This commit is a build-time solution to not generate the traces in the
early config.

The traces are disabled by default, unless the environment STEAMOS_DEBUG
is set. If STEAMOS_DEBUG is set, the early config is built with those
debug traces and built with a delay of 10s before loading the real grub
config.

Note: To re-enable these traces, grub must be re-installed using the
command below:

	$ STEAMOS_DEBUG=1 grub-install

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>