steamos-glx-{backend,driver}: Split out the list of classes into arguments
TL;DR; The steamos-glx-driver now accepts several arguments: one class
per argument.

The script steamos-glx-back gives a list of the GPU classes to the
script steamos-glx-driver. This list is given as a single string
argument.

The elements of that list are separated by a space. The list starts with
a space (as in " ??? nvidia").

This commit updates both scripts to split out the list that is
represented by a single string into several strings where each element
is a class and is represented by a single string.

Behind the hood, the commit uses a bash array that simplifies the code
using both the operator += and the loop for that iterates arrays.

This fixes the (potential?) case where the class nvidia is given as the
last class of a two classes argument (i.e. " ??? nvidia").

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>