steamos-glx-{backend,driver}: Double quote variables and strings
The shell splits out the value of the variables using IFS and then
expands every of its elements using the globing.

For example:

	$ var="The character * is a wildcard"
	$ echo $var
	The character atomic-update chainloader common.mk COPYING.LGPL-2.1 dracut glx gpd-quirks grub Makefile misc offload README.md swap is a wildcard
	              ^---- These values are the content of current directory, printed because of globbing of the character * -----------^
	$ printf $var
	The
	    ^---- The formated string is "The",
	          "character" is the value of the placeholder,
	          "atomic-update" the second,
	          "chainloader" the third,
	          (...)
	          "wildcard" the last

This commit double quotes the variables to prevent them from beeing
splitted and globbed. It is a good habbits to double quote the variables
to prevent from unexpected behaviours due to the word splitting and the
globbing.

Note: The variable module_list is list of modules. The elements of that
list are separated by a space. This commit does not double quote the
variable module_list as the word spilitting is expected here.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>