steamos-glx-driver: Use bash array for module_list
The variable module_list is list of modules. The elements of that list
are separated by a space.

The shell splits out the value of the variables using IFS and then
expands every of its elements using the globing.

This commit uses a bash array to represent that list. The arrays can be
double quoted in its form "${var[@]}". This form prevents the elements
of that variable from beeing splitted and globbed. It is a good habbit
to use bash arrays for lists and to double quote them to prevent from
unexpected behaviours due to the word splitting and the globbing.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>