steamos-glx-driver: Get rid of oIFS
The local variable oIFS is used to store the initial IFS value and to
restore it then.

The IFS value is temporarly updated to add the characters colon and coma
to the IFS in order to read the output of the command dkms status.

	[root@steamos ~]# dkms status nvidia
	nvidia, 440.44, 5.4.10-arch1-1.1, x86_64: installed
	^----^  ^----^  ^--------------^  ^----^  ^-------^
          1       2       3                 4       5

The change of the IFS value lasts for the single intruction that reads
the output of the dkms status only.

This commit reduces the modification of IFS value for the instructions
that matter only; this makes the oIFS variable useless and thus the
variable can be removed safely.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>