steamos-glx-driver: Give the modpath as argument to functions
Both functions get_module_list() and initrd_is_up_to_date() rebuild
internally the path to the module the same way (almost[*]).

	local modpath="/var/lib/dkms/$mod/$ver/$krn/$arc/module";

Both functions use the four parameters (mod, ver, krn, and arc) to
rebuild the path to the modules; it is their only duty.

Both functions are called by the function check_nvidia_dkms_modules();
they are not used elsewhere in the script.

This duplicates the code that makes it more complicated that it should.

This commit factorizes the code by building the path to the module at
one place: in the function check_nvidia_dkms_modules(). Then, the path
to the module is given as parameter to both functions get_module_list()
and initrd_is_up_to_date().

[*]: The function get_module_list() uses uname -r and uname -m if the
paramaters krn and arc are unset. This situation seems to happen if the
command dkms status reports that the module status is added only; this
should not happen if the module is built or is installed.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>