steamos-glx-driver: Do not strip the module extention
The function get_module_list() creates a list of the modules at the
given path. The function strips the kernel extention from the module
names.

The main concern of the extention stripping is Dracut.

The man-page of dracut.conf says:

	add_drivers+=" <kernel modules> "

	Specify a space-separated list of kernel modules to add to the
	initramfs. The kernel modules have to be specified without the
	".ko" suffix[*].

Dealing with the module name that have stripped extention has led to the
commit da6b118 (steamos-glx-driver: Add support for both .ko and .ko.xz
modules). Additional commits such as that one are upcomming if the
distribution changes the compression method in the future; the script
needs to be updated in consequence.

This commit stops to deal with module name that have stripped extention
in its logic. Instead, it strips the module extention whenever it is
needed (i.e. to generate the additional dracut configuration file for
the nvidia driver).

[*]: Local tests show that it is not mandatory to strip the file
extention .ko* (as for dracut version 049). The following dracut
configuration includes the specified modules in the initial ram-disk
image. Let's satisfy the recommentation and strip the file extention
anyway for former versions of dracut that does not support it.

	add_drivers+=" nvidia-drm.ko.xz "
	add_drivers+=" nvidia-modeset.ko.xz "
	add_drivers+=" nvidia-uvm.ko.xz "
	add_drivers+=" nvidia.ko.xz "

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>