steamos-glx-driver: Add a helper to list the missing modules
The helper get_module_list() lists the dkms modules on the system; and
the helper initrd_get_module_list() lists the modules installed in the
image initrd.

This commit introduces the third helper get_missing_module_list() that
lists the modules that are missing from the image initrd. That helper
uses the command comm to compare the two sorted lists of modules and to
output the ones that are unique to one list.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>