steamos-glx-driver: Add a helper to list the up-to-date modules
The helper get_module_list() lists the dkms modules on the system; the
helper initrd_get_module_list() lists the modules installed in the image
initrd; and the helper get_missing_module_list() lists the modules that
are missing from the image initrd.

This commit introduces the fourth helper get_uptodate_module_list() that
lists the modules that are more recent that the image initrd. This
helper simplifies the code the function initrd_is_up_to_date().

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>