steamos-glx-driver: Rework the logic of check_nvidia_dkms_modules
This algorithm in the function check_nvidia_dkms_modules() looks
complicated.

First, it gets the status of the module. Then, it builds the module if
it is not installed already (i.e. the status is added or built); and it
gets the new status of the module.

Note: it does nothing related to dkms if the module is not even present
in the rootfs (noop); in this situation it removes the dracut
configuration related to the module nvidia and exits early with success.

At that point, the status of the module is built at least. The dracut
configuration file is generated to add the modules nvidia to the initial
ram-disk.

The script exits without rebooting if the modules are embedded in the
ram-disk image already and if the image is more recent than the modules.

Otherwise, the script triggers a reboot. There are two paths before the
system reboots:

	- If the modules are installed, it rebuilts the initial ram-disk
	  using dracut.

	- If the modules are built; it installs the module nvidia first
	  and then it rebuilts the initial ram-disk if the ram-disk
	  image is outdated[1].

The chart below describes what has been said:

		dkms status - [(empty)] ---> exit
		|	|
		|	[installed]
		|		  |
		[added|built]	  |
		|		  |
		dkms built	  |
		|		  |
		\________________/
			|
		initramfs up-to-date - [yes] ---> exit
			|
			[no]
			|
		dkms status
		|	|
		|	[built]
		|	      |
		[installed]   |
		|	      |
		dracut        dkms install
		|	      |
		|	      initramfs up-to-date[1]
		|	      |		|
		|	      [no]	[yes]
		|	      |		|
		dracut	      dracut	|
		|	      |		|
		|	      \________/
		|	       |
		\______________/
			|
			v
			reboot[2]

This commit rewrites the function check_nvidia_dkms_modules to make it
more simple.

The first change is about dkms; the function now installs directly the
modules if it is not installed. The module nouveau on Arch Linux[2] is
blacklisted; but this module needs to be installed anyway otherwise no
UI is going to start.

The second change is about the reboot; the script triggers a reboot if
the module is not loaded[2].

[1]: dkms can rebuilt the initial ram-disk (as of the comment).

	# DKMS can rebuild the initrd as part of 'dkms install',  if ever
	# REMAKE_INITRD is set.  The only probem is that we don't know if
	# this option was set (could be added in a conf file by the user).
	# So the only way to know whether we need to rebuild the initrd,
	# AFAICT, is to call initrd_is_up_to_date() a second time. Too bad
	# it's such an expensive operation (lsinitrd under the hood)...

[2]: the final step reboot is *required*: the proprietary module
nvidia-dkms blacklists the open-source module nouveau that causes it to
be unload latter in the boot-sequence.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>