atomupd: Add steamos-udpate-os-plymouth service
TL;DR; This commit adds a bash script that does the glue between
RAUC/casync and Plymouth and adds its systemd unit service.

The SteamOS has its own dedicated systemd target to enter the mode
system updates at start-up. The target is triggered via the script
steamos-update-os and its argument after-reboot; the system has to be
rebooted then.

The systemd unit target steamos-update-os.target pulls the systemd unit
service steamos-update-os.service. It is responsible of starting the
system updates and rebooting the system (to the default mode). It uses
the bash script steamos-update-os that runs steamos-atomupd-client that
pulls a RAUC/casync bundle.

There is no UI to report feedback about the system updates to the user
for now. Even worse, a getty is spawned on tty1 that hides the text
progression of the system-update written to the journal and stops
Plymouth if it is started by the initial ram-disk.

Plymouth has the specific mode system-upgrades that displays the system
updates progression within a progression bar. The progression is updated
using the plymouth client interface.

The RAUC reporting of the update progression is useless if it uses a
casync bundle. It hits the 95% of the update completion in the very
first seconds. RAUC allocates the remaining 5% to the casync process;
casync is the actual system updates and it takes several minutes to
complete. Therefore, the RAUC progression cannot be used as is; the
casync progression should be take into account.

There is an effort in the SteamOS's fork of the project casync to report
some progression. It writes to stdout the seeding and the downloading
chunk progression[1].

RAUC is client/server service. The server part does not forward neither
the stdout nor the stderr of it sub-processes (i.e. casync). Those
traces land to the journal of the service RAUC directly. Thus, the
progression of the casync extraction are available to that place only.

There is some development to do to make the server part of RAUC to
forward both stdout and stderr the client part.

Instead, a reasonable and intermediate solution is to read the journal
of the systemd unit service of RAUC, to extract the progression and to
send it to Plymouth.

This commit introduces the bash script steamos-update-os-plymouth that
does the glue between RAUC/casync and Plymouth. It introduces its unit
service as well.

The bash script is a infinite while loop that read the journal of RAUC.
It never ends unless journald returned; that simplifies the design. Its
goals is to:

	* reset the progress status to 0 if an atomic system update is
	  started or if the service systemd RAUC is beeing started again
	* extract the seed and download progression reported by casync
	  and update the Plymouth progression accordingly

The RAUC casync bundle updates both partitions verity and rootfs. The
measures show of the update of the verity partition takes around 1% of
the total of the system updates. Therefore, the progression of the
verity update is not significant and can be ignored. The update of the
rootfs takes arround 92% of the total update. The last 7% is elapsed in
the post-processing update (regenerating the grub configuration, the
dracut initial ramdisk...). In local, the seeding step is arround 33%
and the downloading step is 67%.

With these numbers and to make the as simple as possible (it is a glue,
and a not-so-dirty hack) the scripts allocates 3 progression slots:

	* 5% for pre-processing: the time elapsed before the atomic
	  update is actually started (systemd services), the download of
	  the bundle (0%), the pre-install script (0%), the casync
	  extraction of the partition verity (1%)
	* 90% for the casync extraction of the rootfs partition
	* 5% for post-processing: the post-install script (7%)

The ratio of seeding/downloading chunks is adjusted to 25%/75% instead
of the of 33%/67% that was measured initially because of the tests were
made on a local network. Latter adjustements should be performed in a
real world situtation.

Also, th script intends to be resistent to the re-start of services and
to the system-update failures.

The systemd service has to depends on the service plymouth-start so it
conflicts with the service getty@tty1 and prevents a login from beeing
spawned to tty1 that causes Plymouth to quit. Furthermore, that service
must be ready before even RAUC and steamos-update-os has been started.
It is started before those two services even if the script is resistent
to start ordering.

[1]: For now, it prints the percentage form 0% to 90%.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>