steamos-glx-driver: Fix the retrieval of the dkms status
The command dkms status reports *all* the statuses for the given module
by version of the kernel installed on the system; one line per kernel
version[1].

The script reads the first status report only. This works if a single
kernel is installed, or if the command dkms starts to report the status
of the module for the version of the kernel that is currently running.

This commits adds the helper get_status() that prints the status of the
module for the version of the kernel that is currently running. If the
module is not even built (i.e. added), it prints the module name, the
version and the status "added" (as dkms).

Note: the helper prints neither the coma nor the colon characters; it
prints the string values directly.

[1]: https://github.com/dell/dkms/blob/v2.8.1/dkms#L2105-L2123

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>