steamos-glx-driver: Add mechanism to prevent reboot in loop
The system may reboot in loop if dpkg cannot build or install the
nvidia-drm driver.

This commit touches a file before rebooting the system. If the system
has to be rebooted after the next boot but the file exists, the script
displays a critical message instead of rebooting the system and exit
failure.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>