steamos-glx-driver: Fallback to driver nouveau if not rebooting
If the driver nvidia is not currently loaded and after the system has
reboot already, a critical message is displayed to the screen saying to
enter the second virtual terminal in order to debug things.

Then, it stops Plymouth causes the display of the systemd logs and does
not start the UI because no there is no graphical driver loaded: the
driver nvidia because of a bug and the driver nouveau because it is
blacklisted by nvidia.

This commit modprobes the driver nouveau if the system reaches this
situation (no nvidia and reboot already). Then, the display manager can
start that makes the debug things easier for people.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>