steamos-glx-driver: Blacklist back nouveau if nvidia is installed
The DKMS drivers are unbuilt and thus are not shipped in the image; the
sources remain in the image so modules can be rebuilt afterward.

The DKMS module nvidia blacklists the module nouveau that prevents it to
be loaded in favour of itself.

As a consequence, there are no graphic backend drivers loaded if this
module is unbuilt and thus is not present in the initial ram-disk.

Therefore, Plymouth displays the three dots on a grey background splash
screen instead of the nice spinner. This is a fallback situation if
there is no DRM device.

The module nouveau was removed from the blacklist of the modules to
prevent that situation. But the module nouveau takes precedence over the
module nvidia, even if it is installed.

As a consequence, the module nouveau must be blacklisted again upon the
module nvidia is installed in the initial ram-disk.

This commit removes the driver nouveau from blacklist if the driver
nvidia is not even built and it re-add the driver nouveau to the
blacklist if the driver nvidia is installed.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>