grub: Remove the disable UUIDs options
These disable options are normally set in the file /etc/default/grub.
It is hard to follow if options are set in multiple places.

This commit removes the forcing of these options since they are set to
the same value already in the file /etc/default/grub shipped by the
package grub.