update-grub: Set the error option
The script update-grub returns success even if the command grub-mkconfig
fails.

This commit sets the option error to the script update-grub so it exits
immediately if a command returns a non-zero status. This avoid to lose
the return status of the command grub-mkconfig if it fails (or any of
the other commands in the script).

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>