steamos-install-steamcl: Do not install if removable media
SteamOS installs its chainloader on every machine when it boots, even if
it is running from a removable media (i.e. /EFI/BOOT/BOOTX86_64.EFI).

This behaviour polutes unnecessarly all the machines on which the
installer is booted.

Even worse, the USB installer replaces the NVRAM entry (if it is more
recent that the copy that is installed) and breaks the boot if the
machine is rebooted on the SteamOS copy without the USB installer.

This commit installs the SteamOS chainloader only if the remobale
bootloader /esp/EFI/BOOT/BOOTX86_64.EFI does not exist.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>