grub: Hit any key to continue booting instead of waiting for a delay
The early config outputs are disabled since commit aa7e4b8 (grub: Remove
traces in early config) and are enabled if the grub image is installed
with the environment variable STEAMOS_DEBUG set.

The outputs are displays for 10 seconds that is not sufisant to debug.

This commit replaces the 10 seconds delay and waits the user to hit a
key instead to continue booting.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>