chainloader: Remove installation of boot service for atomupd target
The service steamos-boot.service is installed for the target
steamos-update-os.target since commit 25e8c23 (atomupd: Remove the
target multi-user.target).

The service never starts as it requires the graphical.target that is
never reached by target steamos-update-os.target (not enabled and
reboot).

This commit removes the installation for steamos-update-os.target.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>