misc: Make steamos-chroot unmount recursively some mountpoints
The helper steamos-chroot prepares the SteamOS environment at a
temporary mount-point, and chroot in there to run a command then.

Commands may mount additionnal mount-points in the chroot environment
and left them mounted (does not mount them after exit).

This leads to the following error:

	# steamos-chroot --partset other -- os-prober
	(...)
	+ umount /tmp/tmp.cdgmb0OEn6/var
	+ umount /tmp/tmp.cdgmb0OEn6/esp
	+ umount /tmp/tmp.cdgmb0OEn6/efi
	+ umount /tmp/tmp.cdgmb0OEn6/tmp
	+ umount /tmp/tmp.cdgmb0OEn6/sys/firmware/efi/efivars
	+ umount /tmp/tmp.cdgmb0OEn6/sys
	umount: /tmp/tmp.cdgmb0OEn6/sys: target is busy.
	+ :
	+ umount /tmp/tmp.cdgmb0OEn6/run
	+ umount /tmp/tmp.cdgmb0OEn6/proc
	+ umount /tmp/tmp.cdgmb0OEn6/dev
	+ umount /tmp/tmp.cdgmb0OEn6
	umount: /tmp/tmp.cdgmb0OEn6: target is busy.
	+ :
	+ rmdir /tmp/tmp.cdgmb0OEn6
	rmdir: failed to remove '/tmp/tmp.cdgmb0OEn6': Device or resource busy

This behaviour is not wanted and it is the fault of the command itself.
But in some situations, it is not; there are some filesystems that are
mounted automatically when they are needed, they are left mounted then.

This is the case of fusectl[1] that is in used by osprober. It is
mounted under the sysfs mount-point /sys.

The command mount shows why the command above could not un-mount the
sysfs in the chroot environment:

	# mount
	(...)
	fusectl on /sys/fs/fuse/connections type fusectl (rw,nosuid,nodev,noexec,relatime)
	fusectl on /tmp/tmp.cdgmb0OEn6/sys/fs/fuse/connections type fusectl (rw,nosuid,nodev,noexec,relatime)

Therefore, the sysfs must be unmounted recursively to make sure all
filesystems are unmounted.

The procfs is another linux filesystem that should also be un-mounted
recursively: the filesystem binfmt_misc[2][3] is typically a filesystem
that can be mounted on top of the procfs mount-point /proc.

This commit unmounts recursively both mountpoints /proc and /sys to make
sure the unmount teardown succeeds and removes the temporary directory.

[1]: https://github.com/systemd/systemd/blob/master/units/sys-fs-fuse-connections.mount
[2]: https://github.com/systemd/systemd/blob/master/units/proc-sys-fs-binfmt_misc.automount
[3]: https://github.com/systemd/systemd/blob/master/units/proc-sys-fs-binfmt_misc.mount

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>