steamos-glx-driver: Fix sed failure on read-only filesystem
The filesystem can be read-only that causes the command sed to on the
file /usr/lib/modprobe.d/nvidia-dkms.conf.

This commit creates and removes the files nvidia-dkms.conf under the
directory /etc/modprobe.d that lives in a filesystem that is writable
instead of sedding the files in /usr/lib/modprobe.d.

Note: The issue failed silently.

Signed-off-by: Gaël PORTAY <gael.portay@collabora.com>