misc: Mount etc even if roothash= is not set
The service etc.mount provides a way to have the directory /etc writable
on a read-only system (i.e. if dm-verity is set up).

There are files that lives there that should not be modified qpunless
you now exactly what you are doing. The overlay file-system preserves
the original files but let you to modify them.

For example, the auto-login file for sddm is present in /etc to set up
auto-login for the user root and start calamares to set up the machine.
This file is deleted after the initial machine setup.

This file is marked in the overlay by the creation of a special file
with the couple major/minor 0/0 in the upper directory. It could be
recreated by the command below:

	$ mknod --mode 0 /var/lib/overlays/etc/upper/sddm.conf.d/calamares-setup.conf c 0 0.

If the etc is not mounted, the auto-login on a writable system is
broken. Even worse, it appears that login is totally broken! But the
investigation is stopped there.

Therefore, it appears that keeping the directory /etc mounted for both
systems read-only and read-write in a sane behaviour to let the system
functional by switching from read-only to read-write, back and forth.

This commit mounts the etc overlay even if in the uneeded situation
where the system is already writable (i.e. if the the roothash= is not
set).